% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suma.R
\name{suma}
\alias{suma}
\title{Suma dos numeros enteros o decimales}
\usage{
suma(x, y)
}
\arguments{
\item{x}{Un numero entero o decimal}

\item{y}{Un numero entero o decimal}
}
\value{
Un numero entero o decimal
}
\description{
Suma dos numeros enteros o decimales
}
\examples{
suma(3, 4)
suma(3.14, 1.3333)
}
