% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiempo.R
\name{what_time}
\alias{what_time}
\title{Title}
\usage{
what_time(language = "es")
}
\arguments{
\item{language}{}
}
\value{
cadena de caracteres con la hora
}
\description{
Title
}
\examples{
what_time()
what_time("en")
}
