% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{what_time}
\alias{what_time}
\title{Decir qué hora es}
\usage{
what_time(language = "es")
}
\arguments{
\item{language}{Idioma, "es" o "en".}
}
\value{
Cadena de caracteres con la hora.
}
\description{
Decir qué hora es
}
\examples{
what_time()
what_time("en")
}
